CCDを修正 無限ループの抜本的な対策をしました。 Epsilonを元にもどしたので、どこかでエラーがでるかもしれません。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2635 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDSphere.cpp
@@ -55,7 +55,7 @@ float CDSphere::GetRadius() {
 	return radius;
 }
 float CDSphere::CalcVolume(){
-	return 4.0f/3.0f*M_PI*radius*radius*radius;
+	return float(4.0/3.0*M_PI*radius*radius*radius);
 }
 Matrix3f CDSphere::CalcMomentOfInertia(){
 	return Matrix3f();
